dev-requirements.txt
cryptography==45.0.4
kubernetes==32.0.1
pillow==11.3.0
playwright==1.53.0
pytest==8.3.5
pytest-playwright==0.7.0
python-keycloak==5.6.0
requests==2.32.4
